Rename top mentions serices
