+++ article/app/controllers/LiveBlogController.scala
@@ -19,7 +19,7 @@ import play.twirl.api.Html
 import renderers.DotcomRenderingService
 import services.CAPILookup
 import services.dotcomponents.DotcomponentsLogger
-import topmentions.TopicService
+import topics.TopicService
 import views.support.RenderOtherStatus
 import scala.concurrent.Future
